用户管理完成

+++ stock_backend/src/main/java/com/cai/stock/vo/resp/PageResult.java
@@ -17,7 +17,7 @@ public class PageResult<T> implements Serializable {
      * 总记录数
      */
     @ApiModelProperty("总记录数")
-    private Long totalRows;
+    private Integer totalRows;
 
     /**
      * 总页数
@@ -52,7 +52,7 @@ public class PageResult<T> implements Serializable {
      * @return
      */
     public PageResult(PageInfo<T> pageInfo) {
-        totalRows = pageInfo.getTotal();
+        totalRows =(int) pageInfo.getTotal();
         totalPages = pageInfo.getPages();
         pageNum = pageInfo.getPageNum();
         pageSize = pageInfo.getPageSize();
